Refactoring: Split a header file to avoid to be longer near future.
